refactor(pages): update pages to use ViewModel pattern and add form utilities
- update shopping page to use useShoppingPageViewModel
- update admin page to use useAdminViewModel
- add use-form hook for form state management
- update shared hooks index for better organization
- improve separation of concerns between UI and business logic
