+++ src/basic/shared/hooks/index.ts
@@ -1,1 +1,2 @@
+export * from './use-form';
 export * from './use-local-storage';
